refactor(theme-default): add `usePrefersColorScheme`
