+++ packages/@vuepress/theme-default/src/client/composables/index.ts
@@ -1,5 +1,6 @@
 export * from './useDarkMode'
 export * from './useNavLink'
+export * from './usePrefersColorScheme'
 export * from './useResolveRouteWithRedirect'
 export * from './useScrollPromise'
 export * from './useSidebarItems'
